+++ examples/stream-payments.php
@@ -3,17 +3,16 @@
 require '../vendor/autoload.php';
 
 use \ZuluCrypto\StellarSdk\Horizon\ApiClient;
-use \ZuluCrypto\StellarSdk\Model\AssetTransferInterface;
+use \ZuluCrypto\StellarSdk\Model\Payment;
 
 $client = ApiClient::newPublicClient();
 
-$client->streamPayments('now', function(AssetTransferInterface $assetTransfer) {
-    printf('[%s] [%s] %s from %s -> %s' . PHP_EOL,
+$client->streamPayments('now', function(Payment $payment) {
+    printf('[%s] %s from %s -> %s' . PHP_EOL,
         (new \DateTime())->format('Y-m-d h:i:sa'),
-        $assetTransfer->getAssetTransferType(),
-        $assetTransfer->getAssetAmount(),
-        $assetTransfer->getFromAccountId(),
-        $assetTransfer->getToAccountId()
+        $payment->getAmount(),
+        $payment->getFromAccountId(),
+        $payment->getToAccountId()
     );
 });