+++ getting-started/06-list-payments.php
@@ -18,7 +18,7 @@ while (true) {
         /** @var $payment \ZuluCrypto\StellarSdk\Model\Operation|\ZuluCrypto\StellarSdk\Model\AssetTransferInterface */
         // If the same cursor shows up twice, we're repeating results and should exit
         if ($payment->getPagingToken() == $currentCursor) break 2;
-
+ 
         printf('[%s] Amount: %s From %s in Tx %s' . PHP_EOL,
             $payment->getAssetTransferType(),
             $payment->getAssetAmount(),